# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Address:
    properties:
      addressLine1:
        type: string
      addressLine2:
        type: string
      city:
        type: string
      createdAt:
        type: string
      fullName:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      phone:
        type: string
      state:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      zipCode:
        type: string
    type: object
...
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parentId:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.CreateOrderInput:
    properties:
      addressId:
        type: integer
      remark:
        type: string
    required:
    - addressId
    type: object
...
  models.Order:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      addressId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      remark:
        type: string
      status:
        type: string
      total:
        type: number
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.OrderItem:
    properties:
      id:
        type: integer
      orderId:
        type: integer
      price:
        type: number
      product:
        $ref: '#/definitions/models.Product'
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      images:
        type: string
      isActive:
        type: boolean
      isFeatured:
        type: boolean
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      stock:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@bycigar.com
    name: API Support
  description: BYCIGAR 电商后台 API 接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BYCIGAR API
  version: "1.0"
paths:
...
  /admin/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件到服务器
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 上传图片
      tags:
      - admin-upload
...
  /orders:
    get:
      consumes:
      - application/json
      description: 获取当前用户的订单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取订单列表
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: 从购物车商品创建订单
      parameters:
      - description: 订单信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定订单的详细信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取订单详情
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: 获取产品列表，支持分页、搜索、分类筛选和排序
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 12
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 分类 slug
        in: query
        name: category
        type: string
      - description: 分类 ID
        in: query
        name: categoryId
        type: integer
      - description: 是否精选
        in: query
        name: featured
        type: string
      - default: createdAt
        description: 排序字段
        in: query
        name: sortBy
        type: string
      - default: desc
        description: 排序方向 (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取产品列表
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取产品详情
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个产品
      tags:
      - products
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"